+++ crates/threshold-signature-server/src/user/tests.rs
@@ -35,6 +35,7 @@ use entropy_protocol::{
 use entropy_shared::{
     HashingAlgorithm, KeyVisibility, OcwMessageDkg, DAVE_VERIFYING_KEY, DEFAULT_VERIFYING_KEY,
     DEFAULT_VERIFYING_KEY_NOT_REGISTERED, DEVICE_KEY_HASH, EVE_VERIFYING_KEY, FERDIE_VERIFYING_KEY,
+    NETWORK_PARENT_KEY,
 };
 use entropy_testing_utils::{
     chain_api::{
@@ -361,6 +362,18 @@ async fn test_sign_tx_no_chain() {
 
     for res in test_user_custom_hash_out_of_bounds {
         assert_eq!(res.unwrap().text().await.unwrap(), "Custom hash choice out of bounds");
+    }
+
+    generic_msg.block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number;
+    generic_msg.signature_verifying_key = NETWORK_PARENT_KEY.0.to_vec();
+    let test_user_sign_with_parent_key = submit_transaction_requests(
+        vec![validator_ips_and_keys[1].clone()],
+        generic_msg.clone(),
+        one,
+    )
+    .await;
+    for res in test_user_sign_with_parent_key {
+        assert_eq!(res.unwrap().text().await.unwrap(), "No signing from parent key");
     }
     clean_tests();
 }
@@ -776,6 +789,81 @@ async fn test_store_share() {
 
 #[tokio::test]
 #[serial]
+async fn test_jumpstart_network() {
+    initialize_test_logger().await;
+    clean_tests();
+
+    let alice = AccountKeyring::Alice;
+
+    let cxt = test_context_stationary().await;
+    let (_validator_ips, _validator_ids, _) =
+        spawn_testing_validators(Some(DEFAULT_VERIFYING_KEY.to_vec()), false, false).await;
+    let api = get_api(&cxt.node_proc.ws_url).await.unwrap();
+    let rpc = get_rpc(&cxt.node_proc.ws_url).await.unwrap();
+
+    let client = reqwest::Client::new();
+
+    let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
+    let validators_info = vec![
+        entropy_shared::ValidatorInfo {
+            ip_address: b"127.0.0.1:3001".to_vec(),
+            x25519_public_key: X25519_PUBLIC_KEYS[0],
+            tss_account: TSS_ACCOUNTS[0].clone().encode(),
+        },
+        entropy_shared::ValidatorInfo {
+            ip_address: b"127.0.0.1:3002".to_vec(),
+            x25519_public_key: X25519_PUBLIC_KEYS[1],
+            tss_account: TSS_ACCOUNTS[1].clone().encode(),
+        },
+    ];
+    let onchain_user_request = OcwMessageDkg {
+        sig_request_accounts: vec![H256::zero().encode()],
+        block_number,
+        validators_info,
+    };
+
+    put_jumpstart_request_on_chain(&api, &rpc, &alice).await;
+
+    run_to_block(&rpc, block_number + 1).await;
+
+    // succeeds
+    let user_registration_response = client
+        .post("http://127.0.0.1:3002/user/new")
+        .body(onchain_user_request.clone().encode())
+        .send()
+        .await
+        .unwrap();
+
+    assert_eq!(user_registration_response.text().await.unwrap(), "");
+    // wait for jump start event check that key exists in kvdb
+    for _ in 0..45 {
+        std::thread::sleep(std::time::Duration::from_millis(1000));
+        let block_hash = rpc.chain_get_block_hash(None).await.unwrap();
+        let events = EventsClient::new(api.clone()).at(block_hash.unwrap()).await.unwrap();
+        let jump_start_event = events.find::<entropy::registry::events::FinishedNetworkJumpStart>();
+        for _event in jump_start_event.flatten() {
+            break;
+        }
+    }
+
+    let get_query = UnsafeQuery::new(hex::encode(H256::zero()), [].to_vec()).to_json();
+    // check get key before registration to see if key gets replaced
+    let response_key = client
+        .post("http://127.0.0.1:3001/unsafe/get")
+        .header("Content-Type", "application/json")
+        .body(get_query.clone())
+        .send()
+        .await
+        .unwrap();
+    // check to make sure keyshare is correct
+    let key_share: Option<KeyShare<KeyParams>> =
+        entropy_kvdb::kv_manager::helpers::deserialize(&response_key.bytes().await.unwrap());
+    assert_eq!(key_share.is_some(), true);
+    clean_tests();
+}
+
+#[tokio::test]
+#[serial]
 async fn test_return_addresses_of_subgroup() {
     initialize_test_logger().await;
 
@@ -1050,6 +1138,18 @@ pub async fn put_register_request_on_chain(
         Static(key_visibility),
         program_instance,
     );
+    submit_transaction(api, rpc, &sig_req_account, &registering_tx, None).await.unwrap();
+}
+
+pub async fn put_jumpstart_request_on_chain(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    sig_req_keyring: &Sr25519Keyring,
+) {
+    let sig_req_account =
+        PairSigner::<EntropyConfig, sp_core::sr25519::Pair>::new(sig_req_keyring.pair());
+
+    let registering_tx = entropy::tx().registry().jump_start_network();
     submit_transaction(api, rpc, &sig_req_account, &registering_tx, None).await.unwrap();
 }